pyproject.toml
[project]
name = "social-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.11.13",
    "ipykernel>=6.29.5",
    "jupyter>=1.1.1",
    "langchain>=0.3.19",
    "loguru>=0.7.3",
    "markdownify>=1.0.0",
    "notebook>=7.3.2",
    "openai>=1.64.0",
    "pandas>=2.2.3",
    "pytest-playwright>=0.7.0",
    "qdrant-client>=1.13.2",
    "scrapfly-sdk[speedups]>=0.8.21",
    "selenium>=4.28.1",
    "tqdm>=4.67.1",
    "xinference>=1.3.0.post2",
]
